+++ operator-pipeline-images/operatorcert/oidc_client.py
